backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
Pillow==11.0.0
python-multipart==0.0.17
google-api-core==2.21.0
